Move landing-page mock data generators out of landing-analytics.tsx
Moved lcg, dailyClicks, toDailyPoints, hourlyPoints, and heatmapData
(all pure, no React) to src/app/lib/landing-mock-data.ts: same split
as org-qr.ts/qr-look.ts elsewhere in this branch, and needed here
too -- exporting heatmapData from the .tsx component file tripped
react-doctor's Fast Refresh check.

Added tests/landing-mock-data.test.ts for heatmapData (full
day/hour coverage, deterministic across calls, weekday-business
hours run hotter than weekend). With coverage it drops off fallow's
flagged list (was 8cyc/90CRAP HIGH).

Also extracted DeltaBadge from StatTile (5cyc -> not flagged) and
RangeTabs from LandingAnalyticsMock (86 -> 70 lines), both already
below fallow's HIGH threshold but flagged in the same file.

Verified: bun run check, worker tsc, bun run test (87 pass),
react-doctor --scope changed (100/100), fallow (heatmapData/StatTile
no longer flagged). Live-verified the landing page's analytics
mockup: range tabs (24h/7d/30d/365d) switch the chart and heatmap
correctly, stat deltas render, no console errors.
