+++ src/app/components/landing-analytics.tsx
@@ -1,68 +1,7 @@
 import { useMemo, useState } from "react";
-import type { HeatmapRow, SeriesPoint } from "@/shared/types";
+import type { SeriesPoint } from "@/shared/types";
 import { AreaChart, BarList, Heatmap } from "./charts";
-
-// Deterministic pseudo-random generator (LCG): the mock data is identical on
-// every visit, so the demo never looks stale or jumps between renders.
-function lcg(seed: number) {
-  return () => {
-    seed = (seed * 1103515245 + 12345) % 2 ** 31;
-    return seed / 2 ** 31;
-  };
-}
-
-/** Daily clicks with a weekly rhythm and a gentle upward trend. */
-function dailyClicks(days: number, base: number, growth: number, seed: number): number[] {
-  const rand = lcg(seed);
-  const out: number[] = [];
-  for (let i = 0; i < days; i++) {
-    const weekly = 1 + 0.3 * Math.sin(((i % 7) / 7) * Math.PI * 2);
-    const trend = 1 + (growth * i) / days;
-    out.push(Math.max(1, Math.round(base * weekly * trend * (0.7 + rand() * 0.6))));
-  }
-  return out;
-}
-
-/** Date-labeled points ending today, mirroring the real stats payload. */
-function toDailyPoints(values: number[]): SeriesPoint[] {
-  const now = Date.now();
-  return values.map((clicks, i) => ({
-    day: new Date(now - (values.length - 1 - i) * 86_400_000).toISOString().slice(0, 10),
-    clicks,
-  }));
-}
-
-/** The last 24 hours, labeled like the real hourly buckets ("… 14:00"). */
-function hourlyPoints(): SeriesPoint[] {
-  const rand = lcg(7);
-  const hourMs = 3_600_000;
-  const start = Math.floor((Date.now() - 23 * hourMs) / hourMs) * hourMs;
-  return Array.from({ length: 24 }, (_, i) => {
-    const d = new Date(start + i * hourMs);
-    const hour = Number(d.toISOString().slice(11, 13));
-    const business = hour >= 8 && hour <= 19;
-    const base = business ? 42 : 8;
-    return {
-      day: `${d.toISOString().slice(0, 10)} ${d.toISOString().slice(11, 13)}:00`,
-      clicks: Math.round(base * (0.6 + rand() * 0.8)),
-    };
-  });
-}
-
-/** Weekday-business-hours heatmap, like a small B2B audience would produce. */
-function heatmapData(): HeatmapRow[] {
-  const rand = lcg(99);
-  const rows: HeatmapRow[] = [];
-  for (let d = 0; d < 7; d++) {
-    for (let h = 0; h < 24; h++) {
-      const business = d < 5 && h >= 8 && h <= 18;
-      const evening = h >= 19 && h <= 22;
-      const base = business ? 26 : evening ? 14 : 3;
-      rows.push({ dayOfWeek: d, hour: h, clicks: Math.round(base * (0.5 + rand())) });
-    }
-  }
-  return rows;
-}
+import { dailyClicks, toDailyPoints, hourlyPoints, heatmapData } from "../lib/landing-mock-data";
 
 const RANGES = [
   { label: "24h", days: 1, bucket: "hour" },
@@ -102,21 +41,51 @@ const DEVICES = [
   { key: "Tablet", clicks: 90 },
 ];
 
+function DeltaBadge({ delta }: { delta: number }) {
+  return (
+    <span
+      className={`tnum mt-1 inline-block text-xs ${delta > 0 ? "text-accent-2" : "text-danger"}`}
+    >
+      {delta > 0 ? "+" : ""}
+      {delta}%
+    </span>
+  );
+}
+
 function StatTile({ label, value, delta }: { label: string; value: number; delta?: number }) {
   return (
     <div className="rounded-lg border border-border bg-bg/40 p-3">
       <p className="truncate text-2xs tracking-wider text-muted uppercase">{label}</p>
       <p className="tnum mt-1 text-xl font-bold">{value.toLocaleString()}</p>
-      {delta != null && delta !== 0 && (
-        <span
-          className={`tnum mt-1 inline-block text-xs ${
-            delta > 0 ? "text-accent-2" : "text-danger"
+      {delta != null && delta !== 0 && <DeltaBadge delta={delta} />}
+    </div>
+  );
+}
+
+function RangeTabs({
+  active,
+  onSelect,
+}: {
+  active: (typeof RANGES)[number]["label"];
+  onSelect: (label: (typeof RANGES)[number]["label"]) => void;
+}) {
+  return (
+    <div className="flex items-center gap-1.5">
+      {RANGES.map((r) => (
+        <button
+          key={r.label}
+          type="button"
+          aria-pressed={active === r.label}
+          onClick={() => onSelect(r.label)}
+          className={`cursor-pointer rounded-md px-2 py-1 text-xs transition-colors ${
+            active === r.label
+              ? "bg-accent text-bg"
+              : "text-muted hover:bg-surface-2 hover:text-text"
           }`}
         >
-          {delta > 0 ? "+" : ""}
-          {delta}%
-        </span>
-      )}
+          {r.label}
+        </button>
+      ))}
     </div>
   );
 }
@@ -153,23 +122,7 @@ export function LandingAnalyticsMock() {
               {active.bucket === "hour" ? "Last 24 hours" : `Last ${active.days} days`}
             </p>
           </div>
-          <div className="flex items-center gap-1.5">
-            {RANGES.map((r) => (
-              <button
-                key={r.label}
-                type="button"
-                aria-pressed={range === r.label}
-                onClick={() => setRange(r.label)}
-                className={`cursor-pointer rounded-md px-2 py-1 text-xs transition-colors ${
-                  range === r.label
-                    ? "bg-accent text-bg"
-                    : "text-muted hover:bg-surface-2 hover:text-text"
-                }`}
-              >
-                {r.label}
-              </button>
-            ))}
-          </div>
+          <RangeTabs active={range} onSelect={setRange} />
         </div>
 
         <div className="grid grid-cols-3 gap-3">
